+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/storage/FileManager.java
@@ -148,37 +148,6 @@ public class FileManager {
     }
 
     /**
-     * Get an array of filenames in the path.
-     *
-     * @param path   The directory path.
-     * @param filter The filter to match file names against, may be {@code null}.
-     * @return An array of filename that doesn't include paths.
-     */
-    @SuppressWarnings("WeakerAccess")
-    @NonNull
-    public static String[] getFilenames(@NonNull String path, @Nullable FilenameFilter filter) {
-        File dir = new File(path);
-        if (dir.exists()) {
-            return dir.list(filter);
-        }
-
-        return new String[0];
-    }
-
-    /**
-     * Get the most recently modified file in the directory specified.
-     *
-     * @param path   The directory path.
-     * @param filter The filter to match file names against, may be {@code null}.
-     * @return The last modified file in the directory matching the specified filter, if any matches. {@code null} otherwise.
-     */
-    @SuppressWarnings("WeakerAccess")
-    @Nullable
-    public static File lastModifiedFile(@NonNull String path, @Nullable FilenameFilter filter) {
-        return lastModifiedFile(new File(path), filter);
-    }
-
-    /**
      * Get the most recently modified file in the directory specified.
      *
      * @param dir    The directory.
@@ -199,7 +168,6 @@ public class FileManager {
                         lastModifiedFile = file;
                     }
                 }
-
                 return lastModifiedFile;
             }
         }